+++ apps/web/src/common/ui/loading/LoadingSpinner.tsx
@@ -1,0 +1,20 @@
+'use client';
+
+import Lottie from '@/common/ui/loading/lottie';
+import animationData from '@/assets/loading.json';
+import * as S from './loading.css';
+
+export interface LoadingSpinnerProps {
+  className?: string;
+}
+
+function LoadingSpinner({ className }: LoadingSpinnerProps) {
+  const containerStyle = className ? className : S.container;
+  return (
+    <div className={containerStyle}>
+      <Lottie animationData={animationData} loop={true} className={S.lottieContainer} />
+    </div>
+  );
+}
+
+export default LoadingSpinner;
