+++ apps/web/src/common/ui/loading/index.tsx
@@ -1,17 +1,7 @@
-import Lottie from '@/common/ui/loading/lottie';
-import animationData from '@/assets/loading.json';
-import * as S from './loading.css';
-interface LoadingSpinnerProps {
-  className?: string;
-}
+import dynamic from 'next/dynamic';
 
-function LoadingSpinner({ className }: LoadingSpinnerProps) {
-  const containerStyle = className ? className : S.container;
-  return (
-    <div className={containerStyle}>
-      <Lottie animationData={animationData} loop={true} className={S.lottieContainer} />
-    </div>
-  );
-}
+const LoadingSpinner = dynamic(() => import('./LoadingSpinner'), {
+  ssr: false,
+});
 
 export default LoadingSpinner;
